+++ test.py
@@ -31,7 +31,10 @@ class ExamplesTest(unittest.TestCase):
         cmd = 'openModel("{}")'.format(package_file)
 
         print("Loading Arduino library")
-        res = cls.dymola.ExecuteCommand(cmd)
+        cls.dymola.ExecuteCommand(cmd)
+
+        print("Setting Compiler")
+        cls.dymola.ExecuteCommand('SetDymolaCompiler("vs", {"CCompiler=MSVC","MSVCDir=C:/Program Files (x86)/Microsoft Visual Studio 14.0/Vc"});')
 
     @classmethod
     def tearDownClass(cls):
@@ -41,18 +44,23 @@ class ExamplesTest(unittest.TestCase):
     def test_examples(self):
 
         examples = [
-            ('AnalogReadSerial', 1, ['arduinoUno.A0.v']),
-            ('BarGraph', 10, ['arduinoUno.A0.v', 'arduinoUno.D2.v', 'arduinoUno.D3.v', 'arduinoUno.D4.v', 'arduinoUno.D5.v', 'arduinoUno.D6.v', 'arduinoUno.D7.v', 'arduinoUno.D8.v', 'arduinoUno.D9.v', 'arduinoUno.D10.v', 'arduinoUno.D11.v']),
-            ('Blink', 10, ['arduinoUno.D13.v']),
-            ('BlinkWithoutDelay', 10, ['arduinoUno.D13.v']),
-            ('Button', 10, ['arduinoUno.D2.v', 'arduinoUno.D13.v']),
-            ('Fade', 10, ['arduinoUno.A0.v', 'arduinoUno.D9.v']),
-            ('Ping', 0.21, ['arduinoUno.D7.v']),
+            ('Arduino.Examples.AnalogReadSerial', 1, ['arduinoUno.A0.v']),
+            ('Arduino.Examples.BarGraph', 10, ['arduinoUno.A0.v', 'arduinoUno.D2.v', 'arduinoUno.D3.v', 'arduinoUno.D4.v', 'arduinoUno.D5.v', 'arduinoUno.D6.v', 'arduinoUno.D7.v', 'arduinoUno.D8.v', 'arduinoUno.D9.v', 'arduinoUno.D10.v', 'arduinoUno.D11.v']),
+            ('Arduino.Examples.Blink', 10, ['arduinoUno.D13.v']),
+            ('Arduino.Examples.BlinkWithoutDelay', 10, ['arduinoUno.D13.v']),
+            ('Arduino.Examples.Button', 10, ['arduinoUno.D2.v', 'arduinoUno.D13.v']),
+            ('Arduino.Examples.Fade', 10, ['arduinoUno.A0.v', 'arduinoUno.D9.v']),
+            ('Arduino.Examples.Ping', 0.21, ['arduinoUno.D7.v']),
+            ('Arduino.Examples.Sweep', 1, ['arduinoUno.D9.v']),
+            ('Arduino.Examples.RobotArm', 15, ['baseServo.flange.phi']),
+            ('Arduino.Firmata.Examples.Blink', 10, ['digitalOutput.value']),
+            ('Arduino.Firmata.Examples.Fade', 10, ['pwmOutput.pinConnector']),
+            ('Arduino.Firmata.Examples.AnalogInput', 10, ['analogInput.y']),
+            ('Arduino.Firmata.Examples.Button', 10, ['digitalInput.y']),
+            ('Arduino.Firmata.Examples.Sweep', 10, ['servo.u']),
         ]
 
-        for name, stop_time, signals in examples:
-
-            model = 'Arduino.Examples.' + name
+        for model, stop_time, signals in examples:
 
             print("Simulating %s" % model)
             success = self.dymola.simulateModel(model, stopTime=stop_time)
@@ -60,7 +68,7 @@ class ExamplesTest(unittest.TestCase):
             self.assertTrue(success, "Simulation failed")
 
             self.dymola.plot(signals)
-            self.dymola.ExportPlotAsImage(name + ".png")
+            self.dymola.ExportPlotAsImage(model + ".png")
 
 
 if __name__ == '__main__':